Autoplay next non-live video after play
When many videos given through command line or playlist mode (from file), the mpv player auto closes when it reach the end of a video and the next in order is then opened by ytdash, enabling creation of custom playlist and/or autoplay all videos found when together with search mode. Only for non-live videos, not live streams.